pyproject.toml
[tool.poetry]
name = "vkr"
version = "0.1.0"
description = "LSTM бейзлайн: предсказание ETA автобусов и оценка аномалий (ВКР)"
readme = "README_BASELINE.md"
authors = ["Author <author@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
pandas = ">=1.3.0"
numpy = ">=1.21.0"
torch = ">=2.0.0"
pytorch-lightning = ">=2.0.0"
tensorboard = ">=2.14.0"
PyYAML = ">=6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

# Ноутбук с картой (datasets/map_track.ipynb): folium требует markupsafe < 3
[tool.poetry.group.notebook.dependencies]
folium = ">=0.14.0"
matplotlib = ">=3.5.0"
markupsafe = ">=2.0,<3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
